requirements.txt
aiocqhttp==1.4.4
beautifulsoup4==4.12.3
lxml==4.9.3
numpy==1.25.2
opencv_python==4.8.0.76
pandas==2.2.2
Pillow==9.5.0
Pillow==10.3.0
pycryptodome==3.19.0
python_Levenshtein==0.25.1
Requests==2.31.0
